requirement.txt
Django==1.10.1
lxml==3.6.4
psycopg2==2.6.2
pycrypto==2.6.1
requests==2.11.1
wheel==0.24.0
